requirements.txt
streamlit
requests
openai
python-dotenv
